Move gtm.tsx to services folder
