+++ apps/store/src/pages/_app.tsx
@@ -3,9 +3,9 @@ import { appWithTranslation } from 'next-i18next'
 import type { AppPropsWithLayout } from 'next/app'
 import Head from 'next/head'
 import { ThemeProvider } from 'ui'
-import { GTMAppScript, useGTMRouteEvents } from '@/lib/gtm'
 import { useApollo } from '@/services/apollo/client'
 import * as Datadog from '@/services/datadog'
+import { GTMAppScript, useGTMRouteEvents } from '@/services/gtm'
 import { ShopSessionProvider } from '@/services/shopSession/ShopSessionContext'
 import { initStoryblok } from '@/services/storyblok/storyblok'
