+++ apps/store/src/pages/_document.tsx
@@ -1,7 +1,7 @@
 import Document, { Head, Html, Main, NextScript } from 'next/document'
 import { getCDNFonts } from 'ui'
-import { GTMBodyScript } from '@/lib/gtm'
 import { getLocaleOrFallback } from '@/lib/l10n/locales'
+import { GTMBodyScript } from '@/services/gtm'
 
 export default class MyDocument extends Document {
   lang(): string {
